refactor: test fixtures
